pyproject.toml
[build-system]
requires = ["poetry-core>=1.2.0b2"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "truthful_counterfactuals"
version = "0.1.0"
description = "Insert Description"
license = "MIT license"
authors = ["Jonas Teufel <jonseb1998@gmail.com>"]
maintainers = ["Jonas Teufel <jonseb1998@gmail.com>"]
readme = "README.rst"
keywords = []
packages = [
    { include = "truthful_counterfactuals" },
    { include = "truthful_counterfactuals/experiments"},
]
include = [
    "truthful_counterfactuals/VERSION",
]

[tool.poetry.scripts]
truthful_counterfactuals = 'truthful_counterfactuals.cli:cli'

[tool.poetry.dependencies]
python = ">=3.8.0,<4.0"
pycomex = ">=0.9.2"
click = ">=7.1.2"
jinja2 = ">=3.0.3"
matplotlib = ">=3.5.3"
numpy = ">=1.23.2"
python-decouple = ">=3.6"
poetry-bumpversion = ">=0.3.0"
rich-click = ">=1.7.0,<=1.8.0"
seaborn = ">=0.11.2"
pandas = ">=2.2.3,<3.0.0"
scikit-learn = ">=1.6.1,<2.0.0"
rdkit = ">=2024.9.0,<2025.0.0"

[tool.poetry.dev-dependencies]
pytest = ">=7.1.3"
poetry-bumpversion = ">=0.3.0"

[tool.poetry_bumpversion.file."truthful_counterfactuals/VERSION"]
search = "{current_version}"
replace = "{new_version}"

[tool.poetry_bumpversion.file."README.rst"]
search = 'version-{current_version}-orange.svg'
replace = 'version-{new_version}-orange.svg'

[tool.poetry_bumpversion.file."truthful_counterfactuals/__init__.py"]
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'